feat(datepicker-range): traduz as literais para os idiomas suportados
Traduz as literais para os idiomas suportados ('pt', 'en', 'es', 'ru')

Fixes MDO-40
