+++ projects/ui/src/lib/components/po-field/po-datepicker-range/po-datepicker-range.literals.ts
@@ -1,0 +1,20 @@
+import { PoDatepickerRangeLiterals } from './interfaces/po-datepicker-range-literals.interface';
+
+export const poDatepickerRangeLiteralsDefault = {
+  en: <PoDatepickerRangeLiterals>{
+    invalidFormat: 'Date in invalid format',
+    startDateGreaterThanEndDate: 'Start date greater than end date'
+  },
+  es: <PoDatepickerRangeLiterals>{
+    invalidFormat: 'Fecha en formato no válido',
+    startDateGreaterThanEndDate: 'Fecha de inicio mayor que fecha final'
+  },
+  pt: <PoDatepickerRangeLiterals>{
+    invalidFormat: 'Data no formato inválido',
+    startDateGreaterThanEndDate: 'Data inicial maior que data final'
+  },
+  ru: <PoDatepickerRangeLiterals>{
+    invalidFormat: 'Дата в неверном формате',
+    startDateGreaterThanEndDate: 'Дата начала больше даты окончания'
+  }
+};
